+++ src/components/Form/Col/Col.tsx
@@ -16,7 +16,7 @@ import FormFieldColorPicker from "@root/components/Field/FormFieldColorPicker";
 import FormFieldDate from "@root/components/Field/FormFieldDate/DateField";
 import FormFieldAddress from "@root/components/Field/FormFieldAddress";
 import FormFieldTable from "@root/components/Field/FormFieldTable";
-import FormFieldTextEditor from "@root/forms/FormFieldTextEditor";
+import FormFieldTextEditor from "@root/components/Field/FormFieldTextEditor";
 import FormFieldAdvancedSelection from "@root/components/Field/FormFieldAdvancedSelection";
 import FormFieldMapCoordinates from "@root/components/Field/FormFieldMapCoordinates";
 import FormFieldImageUpload from "@root/components/Field/FormFieldImageUpload";
